Move videoPlayer out of manager file
